refactor - Name 클래스 생성 및 테스트 코드 작성
